requirements.txt
pandas
sqlalchemy
psycopg2-binary
scikit-learn
numpy
streamlit
plotly
